Implement update op for image pipeline
